% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elo-helpers.R
\name{calculate_MOV}
\alias{calculate_MOV}
\title{Helper function to process elo data into long format}
\usage{
calculate_MOV(elo_diff, margin, J = 2.2)
}
\description{
Helper function to process elo data into long format
}
