% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elo-helpers.R
\name{convert_elo_results}
\alias{convert_elo_results}
\title{Helper function to process elo data into long format}
\usage{
convert_elo_results(results)
}
\description{
Helper function to process elo data into long format
}
