% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elo-helpers.R
\name{find_expected_margin}
\alias{find_expected_margin}
\title{Helper function to process elo data into long format}
\usage{
find_expected_margin(elo_difference, M = 400, B = 0.025)
}
\description{
Helper function to process elo data into long format
}
