% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elo-helpers.R
\name{map_margin_to_outcome}
\alias{map_margin_to_outcome}
\title{Helper function to process elo data into long format}
\usage{
map_margin_to_outcome(margin, A = 0, K = 1, B = 0.05, v = 1, Q = 1,
  C = 1)
}
\description{
Helper function to process elo data into long format
}
