% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plussixelo-package.R
\docType{package}
\name{plussixelo-package}
\alias{plussixelo}
\alias{plussixelo-package}
\title{plussixelo: Tools for running ELO calculations on plussixoneblog.com}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: James Day \email{jamesthomasday@gmail.com}

}
\keyword{internal}
