% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elo-process.R
\name{process_historical_elo}
\alias{process_historical_elo}
\title{X}
\usage{
process_historical_elo(results, init_elo = 1500, HGA = 35, k = 20,
  M = 400, B = 0.025, carryover_weight = 0.6)
}
\arguments{
\item{results}{x}

\item{HGA}{Home ground advantage, in arbitrary units}

\item{M}{Weighting factor for ELO calculations}

\item{stdev}{Standard Deviation of results}
}
\value{
A dataframe of simulated results
}
\description{
\code{process_historical_elo} x
}
\details{
INSERT DESCRIPTION
}
