% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elo-process.R
\name{simulate_season}
\alias{simulate_season}
\title{Simulate a season based on ELO}
\usage{
simulate_season(fixture, team_elo = data.frame(), simulation = 1,
  stdev = 41, HGA = 35, k = 20, M = 400, B = 0.025,
  carryover_weight = 0.6, init_elo = 1500)
}
\arguments{
\item{fixture}{A dataframe containing upcoming matches EXPAND}

\item{team_elo}{A dataframe with each teams current ELO ratings}

\item{simulation}{An optional simulation ID number}

\item{stdev}{Standard Deviation of results}

\item{HGA}{Home ground advantage, in arbitrary units}

\item{M}{Weighting factor for ELO calculations}
}
\value{
A dataframe of simulated results
}
\description{
\code{simulated_season} takes a fixture and simulates remaining games based upon the teams starting ELO.
}
\details{
INSERT DESCRIPTION
}
